Edits Elevator and JumpPad and Adds Timed Floor
Creates new materials for the elevator and jump pad. The jump pad now changes the height of the jump and color of the pad if the player bounces on it. There is now a floor that disappears after a set amount of time. The elevator now has different types for moving up/down or right/left.