Adds a Random Spawner for Parts and Alters Level Parts
Adds a plain floor, replaces the up/down elevator with horizontal floor, replaces sides elevator with a vertical elevator. Adds a spawner that can spawn each floor part for the level. Adds front and back pieces to each floor piece to signify the beginning and end of each part for spawning purposes.